Another follow up 3d9c6e7763c3d257593ba159a18dffc435b5ff06 * Player's Roulette variables aren't saved. * Fixed client crash because of wrong packets sent by Roulette. * Changed item * Added a little check on `char` table for PACKETVER >= 20141016. * Thank @aleos89, @icxbb-xx

+++ src/map/pc.h
@@ -650,7 +650,7 @@ struct map_session_data {
 
 	struct {
 		short stage;
-		short prizeIdx;
+		int8 prizeIdx;
 		short prizeStage;
 		bool claimPrize;
 	} roulette;
